Update minesweeper.py
implementing the use of intersection to check the cells in common between sentences. Finding all the values possible in theses cells and trying them back in the original sentence and discarding values that would lead to impossibilities. adding them to the sentence if only 1 value remained or adding them to a ambiguous_intersection to check later if more then 1 value remained.