minesweeper update
apparently, all bugs were removed.
No more errors nor picking mines as safe locations or safe as mines. still failing two of Cs50 checks, failing the ones where it's possible to figure if a cell is a safe or a mine, but succeeding when there are more then one safe of mine. in one of the cases it's returning set().
need to edit quick_check_ambiguous to delete ambiguous when there is only 1 cell and multiple possible values.
and questioning if it's necessary to reaply quick_check_ambiguous. The Ai was capable of wining the game while still having many cells unexplored, so it'd deductions seems more then good enough.