Project 3 - optimization - crossword - update
Created a queue_2s function that uses two stacks to emulate a queue. Initially felt useful considering the size of some word.txt files,  but on a second look, it will only be used to store variables (the spaces for words in structures.txt), and there aren't many of these, this might've been unnecessary, but still served as a learning experience in creating classes.